+++ src/test/suite/index.ts
@@ -1,5 +1,5 @@
 import * as path from 'path'
-// eslint-disable-next-line node/no-extraneous-import
+// eslint-disable-next-line node/no-unpublished-import
 import * as Mocha from 'mocha'
 // eslint-disable-next-line node/no-extraneous-import
 import * as glob from 'glob'
@@ -14,33 +14,37 @@ export function run(): Promise<void> {
 
   const testsRoot = path.resolve(__dirname, '..')
 
-  return new Promise(function executor(resolve, reject) {
-    glob('**/**.test.js', { cwd: testsRoot }, function callback(
-      err: NodeJS.ErrnoException | null,
-      files: readonly string[]
-    ) {
-      if (err) {
-        return reject(err)
-      }
+  return new Promise<void>(function executor(resolve, reject): void {
+    glob(
+      '**/**.test.js',
+      { cwd: testsRoot },
+      function callback(
+        err: NodeJS.ErrnoException | null,
+        files: readonly string[]
+      ) {
+        if (err) {
+          return reject(err)
+        }
 
-      // Add files to the test suite
-      for (const file of files) {
-        mocha.addFile(path.resolve(testsRoot, file))
-      }
+        // Add files to the test suite
+        for (const file of files) {
+          mocha.addFile(path.resolve(testsRoot, file))
+        }
 
-      try {
-        // Run the mocha test
-        mocha.run(function callback(failures) {
-          if (failures > 0) {
-            reject(new Error(`${failures} tests failed.`))
-          } else {
-            resolve()
-          }
-        })
-      } catch (err) {
-        console.error(err)
-        reject(err)
+        try {
+          // Run the mocha test
+          mocha.run(function callback(failures): void {
+            if (failures > 0) {
+              reject(new Error(`${failures} tests failed.`))
+            } else {
+              resolve()
+            }
+          })
+        } catch (err) {
+          console.error(err)
+          reject(err)
+        }
       }
-    })
+    )
   })
 }