+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to the "typed-css-modules-plugin" extension will be document
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [1.0.0] - 2021-10-17
+
+- Minimal vscode version: `1.40.0` (nodejs 12.x).
+- Change `activationEvents` to detect css languages.
+- Added setting to enable/disable formatting generated files using eslint.
+
+## [0.3.1]
+
+- Fixed yarn global modules path.
+
 ## [0.3.0]
 
 - Now used `sass` module instead `node-sass`